# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Opportunity OS CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-cards:
    get:
      operationId: listBusinessCards
      tags: [businessCards]
      summary: List business cards
      parameters:
        - in: query
          name: processingStatus
          schema:
            type: string
        - in: query
          name: reviewStatus
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of business cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCardListResponse"
    post:
      operationId: createBusinessCard
      tags: [businessCards]
      summary: Upload business card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessCardRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCard"
  /business-cards/{id}:
    get:
      operationId: getBusinessCard
      tags: [businessCards]
      summary: Get business card by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCard"
    put:
      operationId: updateBusinessCard
      tags: [businessCards]
      summary: Update business card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessCardRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCard"
  /business-cards/{id}/parse:
    post:
      operationId: parseBusinessCard
      tags: [businessCards]
      summary: Trigger OCR parsing on a business card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parsed result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCard"
  /business-cards/{id}/approve:
    post:
      operationId: approveBusinessCard
      tags: [businessCards]
      summary: Approve business card and create contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveBusinessCardRequest"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveBusinessCardResponse"
  /business-cards/{id}/reject:
    post:
      operationId: rejectBusinessCard
      tags: [businessCards]
      summary: Reject business card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCard"
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: contactId
          schema:
            type: string
        - in: query
          name: organizationId
          schema:
            type: string
        - in: query
          name: opportunityId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    put:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /reports/activities:
    get:
      operationId: getActivityReport
      tags: [reports]
      summary: Get activity report
      parameters:
        - in: query
          name: days
          schema:
            type: integer
      responses:
        "200":
          description: Activity report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityReport"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Contact:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        organizationId:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        fullName:
          type: string
        title:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        status:
          type: string
        notesText:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        organization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
      required:
        - id
        - workspaceId
        - fullName
        - status
        - createdAt
        - updatedAt
...
    CreateContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        title:
          type: string
        department:
          type: string
        email:
          type: string
        phone:
          type: string
        mobile:
          type: string
        linkedinUrl:
          type: string
        source:
          type: string
        organizationId:
          type: string
        notesText:
          type: string
        tagIds:
          type: array
          items:
            type: string
      required:
        - fullName
...
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        organizationType:
          type: string
        industry:
          type: string
          nullable: true
      required:
        - id
        - name
        - organizationType
    Organization:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        legalName:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        organizationType:
          type: string
        industry:
          type: string
          nullable: true
        subIndustry:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        notesText:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        _count:
          type: object
          nullable: true
          properties:
            contacts:
              type: integer
            opportunities:
              type: integer
      required:
        - id
        - workspaceId
        - name
        - organizationType
        - createdAt
        - updatedAt
...
    CreateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        legalName:
          type: string
        website:
          type: string
        phone:
          type: string
        email:
          type: string
        organizationType:
          type: string
        industry:
          type: string
        subIndustry:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        notesText:
          type: string
        tagIds:
          type: array
          items:
            type: string
      required:
        - name
...
    BusinessCard:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        imageUrlFront:
          type: string
        imageUrlBack:
          type: string
          nullable: true
        rawOcrText:
          type: string
          nullable: true
        parsedJson:
          type: object
          nullable: true
          additionalProperties: true
        processingStatus:
          type: string
        reviewStatus:
          type: string
        linkedContactId:
          type: string
          nullable: true
        linkedOrganizationId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        linkedContact:
          $ref: "#/components/schemas/Contact"
          nullable: true
        linkedOrganization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
      required:
        - id
        - workspaceId
        - imageUrlFront
        - processingStatus
        - reviewStatus
        - createdAt
        - updatedAt
    BusinessCardListResponse:
      type: object
      properties:
        businessCards:
          type: array
          items:
            $ref: "#/components/schemas/BusinessCard"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - businessCards
        - total
        - page
        - limit
    CreateBusinessCardRequest:
      type: object
      properties:
        imageUrlFront:
          type: string
        imageUrlBack:
          type: string
      required:
        - imageUrlFront
    UpdateBusinessCardRequest:
      type: object
      properties:
        parsedJson:
          type: object
          nullable: true
          additionalProperties: true
        reviewStatus:
          type: string
        linkedContactId:
          type: string
        linkedOrganizationId:
          type: string
    ApproveBusinessCardRequest:
      type: object
      properties:
        contactData:
          $ref: "#/components/schemas/CreateContactRequest"
        organizationData:
          $ref: "#/components/schemas/CreateOrganizationRequest"
          nullable: true
        mergeWithContactId:
          type: string
          nullable: true
      required:
        - contactData
    ApproveBusinessCardResponse:
      type: object
      properties:
        businessCard:
          $ref: "#/components/schemas/BusinessCard"
        contact:
          $ref: "#/components/schemas/Contact"
        organization:
          $ref: "#/components/schemas/Organization"
          nullable: true
      required:
        - businessCard
        - contact
...
    Activity:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        type:
          type: string
        subject:
          type: string
        description:
          type: string
          nullable: true
        occurredAt:
          type: string
          format: date-time
        contactId:
          type: string
          nullable: true
        organizationId:
          type: string
          nullable: true
        opportunityId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        contact:
          $ref: "#/components/schemas/Contact"
          nullable: true
        organization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
      required:
        - id
        - workspaceId
        - type
        - subject
        - occurredAt
        - createdAt
    ActivityListResponse:
      type: object
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - activities
        - total
        - page
        - limit
    CreateActivityRequest:
      type: object
      properties:
        type:
          type: string
          enum: [CALL, EMAIL, MEETING, CARD_SCAN, NOTE, FOLLOW_UP, EVENT, INTRO]
        subject:
          type: string
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
        contactId:
          type: string
        organizationId:
          type: string
        opportunityId:
          type: string
      required:
        - type
        - subject
    UpdateActivityRequest:
      type: object
      properties:
        type:
          type: string
        subject:
          type: string
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
...
    Tag:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required:
        - id
        - workspaceId
        - name
...
    ActivityReport:
      type: object
      properties:
        byType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
            required:
              - type
              - count
        byDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
            required:
              - date
              - count
        total:
          type: integer
      required:
        - byType
        - byDay
        - total